messaging/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt
PyJWT==2.10.1
sqlparse==0.5.3
mysqlclient==2.2.4
django-extensions
